+++ assets/snippets/DocLister/lib/DLdebug.class.php
@@ -1,5 +1,6 @@
 <?php
-include_once(MODX_BASE_PATH . 'assets/lib/SqlFormatter.php');
+include_once(MODX_BASE_PATH . 'assets/lib/Formatter/SqlFormatter.php');
+include_once(MODX_BASE_PATH . 'assets/lib/Formatter/HtmlFormatter.php');
 
 class DLdebug
 {
@@ -60,20 +61,23 @@ class DLdebug
             }
         }
     }
-    public function updateMessage($message, $key){
+    public function updateMessage($message, $key, $format = null){
         if (is_scalar($key) && !empty($key) && isset($this->_calcLog[$key])) {
             $this->_calcLog[$key]['msg'] = $message;
+            if(!is_null($format)){
+                $this->_calcLog[$key]['format'] = $format;
+            }
         }
     }
 
-    public function debugEnd($key, $msg = null)
+    public function debugEnd($key, $msg = null, $format = null)
     {
         if (is_scalar($key) && isset($this->_calcLog[$key], $this->_calcLog[$key]['time']) && $this->DocLister->getDebug() > 0) {
             $this->_log[$this->countLog()] = array(
                 'msg' => isset($msg) ? $msg : $this->_calcLog[$key]['msg'],
                 'start' => $this->_calcLog[$key]['start'],
                 'time' => microtime(true) - $this->_calcLog[$key]['time'],
-                'format' => $this->_calcLog[$key]['format']
+                'format' => is_null($format) ? $this->_calcLog[$key]['format'] : $format
             );
             unset($this->_calcLog[$key]['time']);
         }
@@ -115,11 +119,38 @@ class DLdebug
                 $item['start'] = isset($item['start']) ? round(floatval($item['start']), 5) : 0;
 
                 if (isset($item['msg'])) {
-                    if(empty($item['format'])){
-                        $item['msg'] = $this->dumpData($item['msg']);
-                    }else{
-                       $item['msg'] = $this->dumpData(SqlFormatter::format($item['msg']), '', null);
+                    if(is_scalar($item['msg'])){
+                        $item['msg'] = array($item['msg']);
                     }
+                    if(is_scalar($item['format'])){
+                        $item['format'] = array($item['format']);
+                    }
+                    $message = '';
+                    $i = 0;
+                    foreach($item['msg'] as $title => $msg){
+                        $format = isset($item['format'][$i]) ? $item['format'][$i] : null;
+                        switch($format){
+                            case 'sql':{
+                                $msg = $this->dumpData(Formatter\SqlFormatter::format($msg), '', null);
+                                break;
+                            }
+                            case 'html':{
+                                $msg = is_numeric($msg) ? $msg : $this->dumpData(Formatter\HtmlFormatter::format($msg), '', null);
+                                break;
+                            }
+                            default:{
+                                $msg = $this->dumpData($msg);
+                                break;
+                            }
+                        }
+                        if(!empty($title) && !is_numeric($title)){
+                            $message .= $this->DocLister->parseChunk('@CODE:<strong>[+title+]</strong>: [+msg+]<br />', compact('msg', 'title'));
+                        }else{
+                            $message .= $msg;
+                        }
+                        $i++;
+                    }
+                    $item['msg'] = $message;
                 } else {
                     $item['msg'] = '';
                 }